Added another example

+++ type-conversion/main.go
@@ -8,7 +8,14 @@ func main() {
 
 	// speed = speed * force // throws an error - invalid operation: speed * force (mismatched types int and float64)
 
-	// solution
-	speed = speed * int(force)
+	// Attempt 1
+	//speed = speed * int(force)
+	//fmt.Println(speed) // result is diffeent than what we expect
+
+	// Attempt 2
+	//speed = float64(speed) * force // thows an error - cannot use float64(speed) * force (type float64) as type int in assignment
+
+	// Attempt 3
+	speed = int(float64(speed) * force)
 	fmt.Println(speed)
 }
